Solved AoC2020 Day 17 Part 2

+++ aoc2020/aoc2020-kotlin/src/test/kotlin/de/havox_design/aoc2020/day17/Day17Test.kt
@@ -15,7 +15,7 @@ class Day17Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         ConwayCubes(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
@@ -28,7 +28,7 @@ class Day17Test {
         @JvmStatic
         private fun getDataForTestProcessPart2(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2020/day17/day17sample.txt", 848L)
+                Arguments.of("de/havox_design/aoc2020/day17/day17sample.txt", 848)
             )
     }
 }
